(1) Fix Windows compile problems for convex/convex contact (note: you can't use "this" in a constructor's initializer list). (2) Add top level Simbody.h header to be used instead of SimTKsimbody.h, to match new product names.

+++ Simbody/tests/TestCollisionDetectionAlgorithm.cpp
@@ -282,7 +282,7 @@ void testEllipsoidEllipsoid() {
         if (!verifyEllipsoidContact(c, ellipsoid1.getRadii(), ellipsoid2.getRadii(), Vec3(0), Vec3(0), body1.getBodyTransform(state2), body2.getBodyTransform(state2)))
             errorCount++;
     }
-    ASSERT(errorCount < contact.size()/10);
+    ASSERT(errorCount < (int)contact.size()/10);
 }
 
 /**
